Refactor test generation
